requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
Flask-Login==0.6.3
WTForms==3.1.2